Rename packet debug to admin
