+++ src/main/kotlin/one/oktw/galaxy/command/Admin.kt
@@ -1,6 +1,6 @@
 package one.oktw.galaxy.command
 
-import one.oktw.galaxy.command.debug.*
+import one.oktw.galaxy.command.admin.*
 import org.spongepowered.api.Sponge
 import org.spongepowered.api.command.CommandResult
 import org.spongepowered.api.command.CommandSource
@@ -9,7 +9,7 @@ import org.spongepowered.api.command.spec.CommandSpec
 
 class Admin : CommandBase {
     override val spec: CommandSpec = CommandSpec.builder()
-        .permission("oktw.command.debug")
+        .permission("oktw.command.admin")
         .child(Gun().spec, "gun")
         .child(Viewer().spec, "viewer")
         .child(TPX().spec, "tpx")
